requirements.txt
mysql-connector-python
scrython
table2ascii
python-decouple
discord.py
nest-asyncio
SQLAlchemy